backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
pyyaml>=6.0
